完成逐小时天气显示

+++ app/src/main/java/com/wjx/android/weather/model/HourlyData.kt
@@ -62,7 +62,7 @@ data class Humidity(
 
 data class Precipitation(
     val datetime: String,
-    val value: Double
+    val value: Int
 )
 
 data class Pressure(
